Move CLI functionality to own module
