+++ plugins/csv/dacman_csv.py
@@ -6,48 +6,3 @@ import pandas as pd
 from enum import Enum
 import numpy as np
 
-
-from pathlib import Path
-from argparse import ArgumentParser
-import json
-
-
-def get_cli_parser():
-    """
-    Create and setup the parser to process the CLI arguments.
-    """
-    parser = ArgumentParser()
-    # using the `type` kwarg is used to automatically validate the parsed value at the CLI invocation stage
-    # invalid values result in an immediate Argparse exception (i.e. the same as providing invalid flags)
-    parser.add_argument('old', type=Path, help='path to old file')
-    parser.add_argument('new', type=Path, help='path to new file')
-    parser.add_argument('-f', '--format', type=str, required=False, choices=['json'], default='json', help='output format of the diff')
-    parser.add_argument('-c', '--config', required=False, default=None, help='path to configuration file')
-
-    return parser
-
-
-def main():
-    parser = get_cli_parser()
-    args = parser.parse_args()
-
-    old, new = args.old, args.new
-    format_ = args.format
-
-    config = args.config
-
-    if config:
-        differ = TableDiffer.from_config_py(Path(config))
-    else:
-        differ = TableDiffer()
-
-    diff = differ(old, new)
-    diff_record = diff.to_record()
-
-    if format_ == 'json':
-        # TODO move this to appropriate DiffView class
-        output = json.dumps(diff_record, indent=4, separators=(', ', ': '))
-        print(output)
-
-if __name__ == "__main__":
-    main()
