+++ plugins/csv/setup.cfg
@@ -9,7 +9,7 @@ description = Find difference between CSV files
 
 [options.entry_points]
 console_scripts = 
-    dacman-csv = dacman_csv:main
+    dacman-csv = dacman_csv.cli:main
 
 [options]
 packages =
